requirements.txt
python
streamlit
python-dotenv
langchain
langchain-google-genai 
google-generativeai